CEDS-1138 Add Declaration Object Factory Mapper
As part of the new mapping strategy, by using the Java generated model
classes, the `Declaration` instance is saved against a property called
`any` that is an attribute of the `Metadata.java` class.
The type of `any` is `Object.class`, so in order for the serialization
to work properly against the Declaration object we add to either replace
`any` with `Declaration` on the service or simply enter the method in
the `ObjectFactory.java` as part of the schema generation.
We decided to go with the latter.

with @mattclark-zerogravit