CEDS-1598 Parsing XML pointers in DMSREJ messages
This is a script that relies on jq/xq, a tool to manipulate xml and json
files.
It allows us to read the XSD and create a list of pairs of "WCOID" and
"WCO Name". This will be useuful for the product team to understand
which field caused a declaration to be rejected and then present that error and
message to the user.